New and improved skeleton
This version of the code now includes threads and a working thread to
create frog candies. Each of the threads (create_frog, create_escargot,
lucy_and_ethel) can now accept a data structure as input when a thread
is generated with pthread_create(). This data structure is called
CONVEYER_STATUS, and it is defined in the lucy_and_ethel.h file.
Therefore, all of the .cpp files include lucy_and_ethel.h so that they
have access to the initializtion of this data structure.

This version simply creates 100 frogs and stops. Production report is
basically complete. The consumers, semaphors, and conveyor belt queue
data structure still need to be created and implemented.